requirements.txt
annotated-types==0.6.0
argcomplete==3.1.6
black==23.12.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==3.13.0
decli==0.6.1
flake8==6.1.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gspread==5.12.4
httplib2==0.22.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
platformdirs==4.1.0
prompt-toolkit==3.0.36
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
questionary==2.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
termcolor==2.4.0
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
zipp==3.17.0
